Rename parse-body.ts to parse-input.ts to match export name
Completes the round-2 review feedback: the module file now matches
the exported parseInput function name.
