+++ apps/server/src/shared/lib/parse-input.ts
